Rewrite realtime handling
Realtime is now handled in a log-safe way; additionally, the log
contains timestamps for all user input, making it possible to
play back a replay in realtime if we want (or write something like a
log to ttyrec or log to video converter).

Currently, night, moonphase, etc. are updated at the end of each
command, rather than at each user input. The latter is also possible
with this codebase, but leads to some strange timing problems (e.g.
if it changes mid-turn, when do you print the message?)

Moonphase and Friday 13th have't been tested (because it isn't a
weird moon or a Friday 13th today on 6 April 2014).

This also exposed a really insidious bug in memfile.c, fixed in this
patch. I'm glad I decided to use subsecond accuracy, or it might not
have been noticed for another 24 years :-)

+++ libnethack/src/mkroom.c
@@ -1,5 +1,5 @@
 /* vim:set cin ft=c sw=4 sts=4 ts=8 et ai cino=Ls\:0t0(0 : -*- mode:c;fill-column:80;tab-width:8;c-basic-offset:4;indent-tabs-mode:nil;c-file-style:"k&r" -*-*/
-/* Last modified by Alex Smith, 2014-04-05 */
+/* Last modified by Alex Smith, 2014-04-06 */
 /* Copyright (c) Stichting Mathematisch Centrum, Amsterdam, 1985. */
 /* NetHack may be freely redistributed.  See license for details. */
 
@@ -429,7 +429,7 @@ antholemon(const d_level * dlev)
     int mtyp;
 
     /* Same monsters within a level, different ones between levels */
-    switch ((level_difficulty(dlev) + ((long)u.ubirthday)) % 3) {
+    switch ((level_difficulty(dlev) + ((unsigned)u.ubirthday)) % 3) {
     default:
         mtyp = PM_GIANT_ANT;
         break;
